backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
groq
requests
google-api-python-client
psycopg[binary]
